# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation by Kang Fakhrur",
        "title": "HRIS Data Warehouse",
        "contact": {},
        "version": "V1.2412.081710"
    },
    "paths": {
...
        "/v1/dashboard/barchart-employee-per-citizen-desc": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Bar Chart Per Citizen Description",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Bar Chart Per Citizen Description",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/dashboard/barchart-employee-per-department": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Bar Chart Per Department",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Bar Chart Per Department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gender (string) (m/f)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/dashboard/barchart-employee-per-gender": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Bar Chart Per Department",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Bar Chart Per Department",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/dashboard/barchart-employee-per-race-desc": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Bar Chart Per Race Description",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Bar Chart Per Race Description",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/dashboard/barchart-employee-per-recruitment-source": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Bar Chart Per Recruitment Source",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Bar Chart Per Recruitment Source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/v1/dashboard/score-card": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Score Card",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Score Card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string) (m/f)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "JwtToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}